+++ R/Hurdleztnegbin.R
@@ -768,28 +768,32 @@ Hurdleztnegbin <- function(nSim = 1000, epsSim = 1e-8, eimStep = 6,
   }
   
   getStart <- expression(
-    if (!is.null(controlMethod$start)) {
-      start <- controlMethod$start
-    } else {
+    if (method == "IRLS") {
+      etaStart <- cbind(
+        log(observed),
+        abs((var(observed) / mean(observed) - 1) / mean(observed)) + .1,
+        (sizeObserved * (observed == 1) + .5) / (sizeObserved + 1)
+      ) + offset
+    } else if (method == "optim") {
       init <- c(
         family$links[[1]](mean(observed)),
         family$links[[2]](abs((var(observed) / mean(observed) - 1) / mean(observed)) + .1),
         family$links[[3]](mean(observed == 1) + .01)
       )
       if (attr(terms, "intercept")) {
-        start <- c(init[1], rep(0, attr(Xvlm, "hwm")[1] - 1))
+        coefStart <- c(init[1], rep(0, attr(Xvlm, "hwm")[1] - 1))
       } else {
-        start <- rep(init[1] / attr(Xvlm, "hwm")[1], attr(Xvlm, "hwm")[1])
+        coefStart <- rep(init[1] / attr(Xvlm, "hwm")[1], attr(Xvlm, "hwm")[1])
       }
       if ("(Intercept):alpha" %in% colnames(Xvlm)) {
-        start <- c(start, init[2], rep(0, attr(Xvlm, "hwm")[2] - 1))
+        coefStart <- c(coefStart, init[2], rep(0, attr(Xvlm, "hwm")[2] - 1))
       } else {
-        start <- c(start, rep(init[2] / attr(Xvlm, "hwm")[2], attr(Xvlm, "hwm")[2]))
+        coefStart <- c(coefStart, rep(init[2] / attr(Xvlm, "hwm")[2], attr(Xvlm, "hwm")[2]))
       }
       if ("(Intercept):pi" %in% colnames(Xvlm)) {
-        start <- c(start, init[3], rep(0, attr(Xvlm, "hwm")[3] - 1))
+        coefStart <- c(coefStart, init[3], rep(0, attr(Xvlm, "hwm")[3] - 1))
       } else {
-        start <- c(start, rep(init[3] / attr(Xvlm, "hwm")[3], attr(Xvlm, "hwm")[3]))
+        coefStart <- c(coefStart, rep(init[3] / attr(Xvlm, "hwm")[3], attr(Xvlm, "hwm")[3]))
       }
     }
   )