+++ R/zotnegbin.R
@@ -485,22 +485,26 @@ zotnegbin <- function(nSim = 1000, epsSim = 1e-8, eimStep = 6,
   }
   
   getStart <- expression(
-    if (!is.null(controlMethod$start)) {
-      start <- controlMethod$start
-    } else {
+    if (method == "IRLS") {
+      etaStart <- cbind(
+        log(observed),
+        abs((var(observed) / mean(observed) - 1) / mean(observed)) + .1
+      ) + offset
+      etaStart <- etaStart[(observed > 1), , drop = FALSE]
+    } else if (method == "optim") {
       init <- c(
         family$links[[1]](mean(observed)),
-        family$links[[2]](abs((var(observed) / mean(observed) - 1) / mean(observed)))
+        family$links[[2]](abs((var(observed) / mean(observed) - 1) / mean(observed)) + .1)
       )
       if (attr(terms, "intercept")) {
-        start <- c(init[1], rep(0, attr(Xvlm, "hwm")[1] - 1))
+        coefStart <- c(init[1], rep(0, attr(Xvlm, "hwm")[1] - 1))
       } else {
-        start <- rep(init[1] / attr(Xvlm, "hwm")[1], attr(Xvlm, "hwm")[1])
+        coefStart <- rep(init[1] / attr(Xvlm, "hwm")[1], attr(Xvlm, "hwm")[1])
       }
       if ("(Intercept):alpha" %in% colnames(Xvlm)) {
-        start <- c(start, init[2], rep(0, attr(Xvlm, "hwm")[2] - 1))
+        coefStart <- c(coefStart, init[2], rep(0, attr(Xvlm, "hwm")[2] - 1))
       } else {
-        start <- c(start, rep(init[2] / attr(Xvlm, "hwm")[2], attr(Xvlm, "hwm")[2]))
+        coefStart <- c(coefStart, rep(init[2] / attr(Xvlm, "hwm")[2], attr(Xvlm, "hwm")[2]))
       }
     }
   )